refactor: 💡 extract common authenticator bits into base class
