+++ addons/auth/tests/dummy/app/authenticators/base.js
@@ -1,0 +1,11 @@
+import BaseAuthenticator from 'auth/authenticators/base';
+
+export default class ApplicationBaseAuthenticator extends BaseAuthenticator {
+  buildTokenValidationEndpointURL(id) {
+    return `/validate/${id}`;
+  }
+
+  buildDeauthEndpointURL() {
+    return '/deauthenticate';
+  }
+}
